refactor: move the DotEnvOptions to EnvHelper
